fix: address PR review findings (batch 14 — typed Stage, posture cast, memgraph errors)
- Add type Stage string to authoring/stages.go for compile-time safety (spgr-cmg.19)
- Update passes.go, prompts.go, authoring_handler.go to use typed Stage keys
- Remove unnecessary float64() casts in DetectPosture (spgr-cmg.56)
- Fix swallowed rec.Get() errors in AmendSpec (spgr-cmg.3)
- Fix inaccurate StoreDecomposeOutput doc comment about MERGE (spgr-cmg.54)

+++ internal/authoring/posture.go
@@ -28,10 +28,10 @@ func DetectPosture(messages []string) specv1.Posture {
 	}
 	avg := float64(total) / float64(len(messages))
 
-	if avg < float64(driveThreshold) {
+	if avg < driveThreshold {
 		return specv1.Posture_POSTURE_DRIVE
 	}
-	if avg > float64(supportThreshold) {
+	if avg > supportThreshold {
 		return specv1.Posture_POSTURE_SUPPORT
 	}
 	return specv1.Posture_POSTURE_PARTNER
